+++ application/command-composition/src/main/kotlin/org/occurrent/application/composition/command/CompositionExtensions.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Johan Haleby
+ * Copyright 2021 Johan Haleby
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,6 +18,7 @@ package org.occurrent.application.composition.command
 
 import java.util.function.Function
 
+// Sequence Composition
 
 /**
  * Compose two commands using infix notation. The resulting command will be executed atomically in the event store.
@@ -33,18 +34,72 @@ import java.util.function.Function
 infix fun <T> ((Sequence<T>) -> Sequence<T>).andThen(anotherCommand: (Sequence<T>) -> Sequence<T>): (Sequence<T>) -> Sequence<T> =
     composeCommands(this, anotherCommand)
 
+/**
+ * Compose multiple commands
+ */
 fun <T> composeCommands(
     firstCommand: (Sequence<T>) -> Sequence<T>,
     secondCommand: (Sequence<T>) -> Sequence<T>,
     vararg additionalCommands: (Sequence<T>) -> Sequence<T>
 ): ((Sequence<T>) -> Sequence<T>) {
-    val commands = sequenceOf(firstCommand, secondCommand, *additionalCommands).map { it.toMutableListCommand() }.toList()
+    return composeCommands(sequenceOf(firstCommand, secondCommand, *additionalCommands))
+}
+
+/**
+ * Compose a sequence of commands
+ */
+fun <T> composeCommands(commands: Sequence<(Sequence<T>) -> Sequence<T>>): (Sequence<T>) -> Sequence<T> {
+    val listCommands = commands.map { it.toMutableListCommandFromSequence() }.toList()
     return { events ->
-        ListCommandComposition.composeCommands(commands).apply(events.toList()).asSequence()
+        ListCommandComposition.composeCommands(listCommands).apply(events.toList()).asSequence()
     }
 }
 
-private fun <T> ((Sequence<T>) -> Sequence<T>).toMutableListCommand(): Function<MutableList<T>, MutableList<T>> =
+// List Composition
+/**
+ * Compose two commands using infix notation. The resulting command will be executed atomically in the event store.
+ * For example:
+ * ```kotlin
+ * val cmd1 : (List<DomainEvent>) -> List<DomainEvent> = ..
+ * val cmd2 : (List<DomainEvent>) -> List<DomainEvent> = ..
+ * applicationService.execute("streamId", cmd1 andThen cmd2)
+ * ```
+ *
+ * @param anotherCommand The other command to run after this one.
+ */
+@JvmName("andThenList")
+infix fun <T> ((List<T>) -> List<T>).andThen(anotherCommand: (List<T>) -> List<T>): (List<T>) -> List<T> =
+    composeCommands(this, anotherCommand)
+
+/**
+ * Compose multiple commands
+ */
+@JvmName("commandsListCommands")
+fun <T> composeCommands(
+    firstCommand: (List<T>) -> List<T>,
+    secondCommand: (List<T>) -> List<T>,
+    vararg additionalCommands: (List<T>) -> List<T>
+): ((List<T>) -> List<T>) {
+    return composeCommands(listOf(firstCommand, secondCommand, *additionalCommands))
+}
+
+/**
+ * Compose a sequence of commands
+ */
+fun <T> composeCommands(commands: List<(List<T>) -> List<T>>): (List<T>) -> List<T> {
+    val listCommands = commands.map { it.toMutableListCommandFromList() }
+    return { events ->
+        ListCommandComposition.composeCommands(listCommands).apply(events.toList())
+    }
+}
+
+// Helpers
+private fun <T> ((Sequence<T>) -> Sequence<T>).toMutableListCommandFromSequence(): Function<MutableList<T>, MutableList<T>> =
     Function<MutableList<T>, MutableList<T>> { events ->
         this(events.asSequence()).toMutableList()
+    }
+
+private fun <T> ((List<T>) -> List<T>).toMutableListCommandFromList(): Function<MutableList<T>, MutableList<T>> =
+    Function<MutableList<T>, MutableList<T>> { events ->
+        this(events).toMutableList()
     }